trabajando en simulacion

+++ src/components/EditForm.jsx
@@ -12,7 +12,7 @@ const EditForm = ({ product, onClose, onSaveEditData }) => {
     const handleSubmit = async () => {
         try {
             const response = await axios.put(`${process.env.REACT_APP_BASE_URL}/products/${editedProduct.objectId}`,
-                { bank: editedProduct.bank, name: editedProduct.name, type: editedProduct.type, apr: parseInt(editedProduct.apr) });
+                { bank: editedProduct.bank, name: editedProduct.name, type: editedProduct.type, apr: parseFloat(editedProduct.apr) });
             console.log(response.data);
             onSaveEditData(response.data);
             onClose(); // close the form upon successful edit
